Security: fix CodeQL cpp/potentially-dangerous-function and integer-overflow alerts
Thread-safe localtime (cpp/potentially-dangerous-function):
- TimeFunctions.cpp: replace localtime() with localtime_r/localtime_s wrapper
- Genome_genomeGenerate.cpp: remove dead asctime(localtime(...)) assignments
  whose results were never used (log lines used timeMonthDayTime() directly)
- Genome_genomeLoad.cpp: replace inline asctime(localtime(...)) with
  timeMonthDayTime() which already provides a thread-safe formatted string

Integer multiplication before cast (cpp/integer-multiplication-cast-to-long):
- ReadAlign.cpp: (uint64_t)P.runRNGseed*(iChunk+1) — was int*int
- SoloFeature_collapseUMIall.cpp: (uint64_t)nGenes*countMatStride — was uint32*uint32
- SoloFeature_emptyDrops_CR.cpp: (uint64_t)ig*stride — was uint32*uint32

+++ source/SoloFeature_collapseUMIall.cpp
@@ -91,8 +91,8 @@ void SoloFeature::collapseUMIperCB(uint32 iCB, vector<uint32> &umiArray, vector<
     
     vector<unordered_map <uintUMI,uintUMI>> umiCorrected(nGenes);
 
-    if (countCellGeneUMI.size() < countCellGeneUMIindex[iCB] + nGenes*countMatStride)
-        countCellGeneUMI.resize((countCellGeneUMI.size() + nGenes*countMatStride )*2);//allocated vector too small
+    if (countCellGeneUMI.size() < countCellGeneUMIindex[iCB] + (uint64_t)nGenes*countMatStride)
+        countCellGeneUMI.resize((countCellGeneUMI.size() + (uint64_t)nGenes*countMatStride)*2);//allocated vector too small
     
     nGenePerCB[iCB]=0;
     nUMIperCB[iCB]=0;
